requirements.txt
asttokens==3.0.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backtrader==1.9.78.123
beautifulsoup4==4.13.3
bs4==0.0.2
bselib==0.0.5
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.2.2
constantly==15.1.0
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
fonttools==4.55.8
frozendict==2.4.6
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.6.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
logzero==1.7.0
lxml==5.3.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
multitasking==0.0.11
nest-asyncio==1.6.0
nselib==1.5
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
Pygments==2.19.1
pyotp==2.8.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pyzmq==26.2.1
rdflib==6.3.2
rdflib-jsonld==0.6.2
requests==2.31.0
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
simplejson==3.19.1
six==1.16.0
smartapi-python==1.4.8
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2025.1
urllib3==2.0.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
yfinance==0.2.52
zope.interface==6.0
